Fix cli execution path under the project
